+++ src/bacnet/basic/object/lc.c
@@ -244,8 +244,8 @@ bool Load_Control_Object_Name(
                 characterstring_init_ansi(object_name, pObject->Object_Name);
         } else {
             snprintf(
-                name_text, sizeof(name_text), "LOAD_CONTROL-%u",
-                object_instance);
+                name_text, sizeof(name_text), "LOAD_CONTROL-%lu",
+                (unsigned long)object_instance);
             status = characterstring_init_ansi(object_name, name_text);
         }
     }
